### 2024-08-06 #7078 @crvernon
:wave: @Stat-Cook - while I am getting you an editor assigned, could you correct the ORICD that is causing the paper compile issue above?  Thanks.
