### 2024-08-15 #7078 @JBorrow
@crvernon unfortunately I don't have any R experience, and my community doesn't use it, so I would not be the best choice to edit this package. I would be happy to take on a different package in different language, though (python, C, golang, fortran, javascript, ...)
