### 2024-08-16 #7078 @RMeli
@crvernon I also have no experience with R at all, and I'll soon go on holidays so I think it would be better to find another editor. But please let me know if you can't find anyone and I'll step in when I'll be back from holidays.
