### 2024-08-28 #7078 @spholmes
Hi @PatrickRWright
Would you be able to review this JOSS submission for us?
Many Thanks
Susan
