### 2024-08-28 #7078 @spholmes
Hello @peerherholz
As an author of a paper on data in JOSS, we thought you might be able to help us out with a
review of this submission on the R package Deident.
Would you be able to help us out here?
Thanks
Susan
