### 2024-08-29 #7078 @PatrickRWright
What are the responsibilities and what is the timeline?

I can really only make technical comments even though I understand, and agree with the general need of such methods. Off the top of my head I was wondering if it may be possible to  show the code coverage badge for the package and why the authors decided not to submit to CRAN.
