### 2024-08-29 #7078 @spholmes
> What are the responsibilities and what is the timeline?
>
> I can really only make technical comments even though I understand, and agree with the general need for such methods. Off the top of my head I was wondering if it may be possible to show the code coverage badge for the package and why the authors decided not to submit to CRAN.

Thanks @PatrickRWright , it is valid to ask the CRAN question and there is a quick checklist the reviewers go trhough, it is public on the github review issue, as was the one for your paper:
https://github.com/openjournals/joss-reviews/issues/2816

We like to get finished with the  review  after a few weeks if possible, the authors respond to the questions as they come up. The code is available and statistics for the code coverage and type are automatically generated byt the github bot.

Let me know if you have other questions, and thanks fore your prompt reply.
Susan
