### 2024-08-29 #7078 @PatrickRWright
1. [DESCRIPTION](https://github.com/Stat-Cook/deident/blob/master/DESCRIPTION) file needs to be adjusted with actual information. It still contains placeholder info.
2. [babyname_sample.rda](https://github.com/Stat-Cook/deident/blob/master/data/babyname_sample.rda) has no trace. i.e. how was it produced. Also it seems to be potentially unused. Maybe its better removed from the package? --> I will not install a package from GItHub with an rda for which I can't find a clear trace.
3. I suggest adding a NEWS.md file to be able to track a sensible changelog.

All above are basic checks for R Packages. I thus suggest going over best practices here https://r-pkgs.org/.
