+++ NEWS.md
@@ -1,0 +1,3 @@
+# deident (development version)
+  
+* Implemented a suite of methods for PID removal.
