+++ doc/REFS.bib
@@ -1,0 +1,6 @@
+@misc{
+  EU_2016, 
+  title = {Regulation (EU) 2016/679 of the European Parliament and of the Council of 27 April 2016 on the protection of natural persons with regard to the processing of personal data and on the free movement of such data (United Kingdom General Data Protection Regulation). Chapter 1; Article 4: Definitions’ Available at: https://www.legislation.gov.uk/eur/2016/679 (accessed on 1st August 2024)},
+  year = 2016,
+  url = https://www.legislation.gov.uk/eur/2016/679
+}
