### 2024-08-29 acaaa07
Remove vestigial documentation
