### 2024-08-29 f51c3f5
Remove vestigial documentation
