+++ R/data.R
@@ -1,23 +1,3 @@
-#' Babyname prevalence for 2017 and 2018
-#'
-#'
-#' Subsample of the `babynames` data set found in
-#' \href{https://github.com/hadley/babynames}{`babynames`}
-#' supplied for testing the package.
-#'
-#'
-#' @format  A data frame with 65,448 rows and 5 columns:
-#' \describe{
-#'   \item{year}{Calendar year considered}
-#'   \item{sex}{Assigned at birth gender - M: male, F: female}
-#'   \item{name}{Popular baby name in given year}
-#'   \item{n}{Frequency to which `name` was used in `year`}
-#'   \item{prop}{Proportion of all baby's born in `year` with `name`}
-#'
-#' }
-#'
-"babyname_sample"
-
 #' Synthetic data set listing daily shift pattern for fictitious employees
 #'
 #' A synthetic data set intended to demonstrate the design and application of a
