### 2024-08-29 #7078 @Stat-Cook
@PatrickRWright thanks for the quick notes.
1. Fixed - sorry overlooked updating this.
2. Thanks for noticing this, old version of examples/ vignettes used this file.   The file and all references in documentation have been removed.
3. NEWS.md has been added.

Currently re-reviewing the package practices but please say if anything else comes to mind.
