### 2024-08-29 #7157 @editorialbot
[REVIEW]: Deident: An R package for data anonymization 1
**Submitting author:** @Stat-Cook (<a href="http://orcid.org/0000-0003-3343-8271">Robert Cook</a>)
**Repository:** https://github.com/Stat-Cook/deident
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.1
