### 2024-08-29 #7078 @spholmes
@Stat-Cook : We have started a new thread with the reviewing issue https://github.com/openjournals/joss-reviews/issues/7157.
