### 2024-08-29 #7157 @spholmes
Hi @Stat-Cook:
The paper format is a bit off, could you put the references in their own section, as it stands they appear in the acknowledgements.
Thanks
Susan
