### 2024-08-29 #7157 @Stat-Cook
@spholmes  Sure - is there a github command to produce the article proof?  Might take a couple of generations.
