### 2024-08-29 4431ac7
Add 'References' section
