+++ paper.Rmd
@@ -115,3 +115,5 @@ The `deident` toolbox was developed for applications in the NuRS and AmreS resea
 
 The development of `deident` was part of the he NuRS and AmreS projects funded by the Health Foundation.
 
+## References
+
