### 2024-08-29 #7157 @PatrickRWright
"Contribution and authorship": I can't judge the contribution of the authors aside from @Stat-Cook since they do not appear in the commit history of the software. I suggest stating individual contributions a dedicated section of the paper.
