### 2024-08-29 #7157 @PatrickRWright
"Worked Example" link here is broken: https://raw.githubusercontent.com/Stat-Cook/deident/master/README.md
