### 2024-08-29 #7157 @PatrickRWright
"Research in the discipline of health data is of increasing interest due to the perception that artificial intelligence (AI) and machine learning (ML) techniques have the promise to bridge gaps in the large, resource-limited sector."
--> I find this sentence too restrictive. "Research in the discipline of health data" has likely always been somewhat equally interesting. I don't see that people care more about health data today (where we have new AI tools) than they may have ten years ago. The current advancements in AI and ML are just adding additional facets to the analytical portfolio.
