### 2024-08-29 #7157 @Stat-Cook
> "Worked Example" link here is broken: https://raw.githubusercontent.com/Stat-Cook/deident/master/README.md

Should now be fixed.
