### 2024-08-29 #7157 @PatrickRWright
It would be nice to have vignette-stlye examples fir all mentioned de-identification methods.
