### 2024-08-29 #7157 @PatrickRWright
"Current applications" -> Is there any citable output from "NuRS and AmreS research projects" that could be mentioned here?
