### 2024-08-29 #7157 @PatrickRWright
Are there really no other R packages attempting this? I do not know but find it hard to believe.
