### 2024-08-29 #7157 @Stat-Cook
> "Current applications" -> Is there any citable output from "NuRS and AmreS research projects" that could be mentioned here?

@PatrickRWright  Under review, unfortunately. Happy to add reference to them once published
