### 2024-08-29 #7157 @PatrickRWright
> > "Current applications" -> Is there any citable output from "NuRS and AmreS research projects" that could be mentioned here?
>
> @PatrickRWright Under review, unfortunately. Happy to add reference to them once published

Is it possible to add working titles and indicate "under review" status?
