### 2024-08-29 #7157 @Stat-Cook
> Are there really no other R packages attempting this? I do not know but find it hard to believe.

I'm yet to find one - one of the core issues is the stateful behavior of the R6 implementation which is deeply non-functional and hence 'not R'.  The only toolsets I've come across have been proprietary.
