### 2024-08-29 #7157 @spholmes
> Are there really no other R packages attempting this? I do not know but find it hard to believe.

There are a few mentoned here:
https://kamran-afzali.github.io/posts/2021-02-14/De-Identification_R.html
