### 2024-08-29 #7157 @PatrickRWright
At this stage I think there are a few things to do so I will forgo further comments until these things have been addressed. I'll point new things out via Issues in https://github.com/Stat-Cook/deident to make it easier.
