### 2024-08-29 25738b4
Remove '.data.frame' from method calls.
