+++ paper.Rmd
@@ -68,7 +68,7 @@ The simplest use case is  transforming one variable via a single method:
 ``` {r, eval=F}
 pipeline <- deident(babynames, "psudonymize", name)
 
-apply_deident.data.frame(babynames, pipeline)
+apply_deident(babynames, pipeline)
 
 ```
 
@@ -77,7 +77,7 @@ The same method can be applied to multiple variables by adding the variable name
 ``` {r, eval=F}
 pipeline2 <- deident(babynames, "psudonymize", name, sex)
 
-apply_deident.data.frame(babynames, pipeline2)
+apply_deident(babynames, pipeline2)
 ```
 
 
@@ -101,7 +101,7 @@ blur <- NumericBlurer$new(cuts=c(0, 10, 100, 1000, 10000))
 pipeline3 <- deident(babynames, "psudonymize", name, sex) |>
   deident(blur, n)
 
-apply_deident.data.frame(babynames, pipeline3)
+apply_deident(babynames, pipeline3)
 ```
 
 An in depth example can be found in the `Worked Example` vigentte, while 
