+++ knit_pub.R
@@ -4,3 +4,4 @@ preamble <- readr::read_lines("preamble.txt")
 
 c(preamble, readr::read_lines("paper.md")) |>
   readr::write_lines("paper.md")
+
