### 2024-08-29 Stat-Cook/deident#5 @Stat-Cook
JOSS Review - add comparisons
Add section to paper comparing utility of this package to those listed at https://kamran-afzali.github.io/posts/2021-02-14/De-Identification_R.html
