### 2024-08-29 Stat-Cook/deident#6 @Stat-Cook
JOSS review - extra vignettes
Can we have a vignette for each method?

* ~add_blur~
* ~add_encrypt~
* ~add_perturb~
* ~add_shuffle (including add_group/ add_ungroup)~
