+++ paper.bib
@@ -76,3 +76,47 @@ urldate = {2023-07-08}
   pages={1--9},
   year={2020}
 }
+
+@Manual{
+    digest2021,
+    title = {digest: Create Compact Hash Digests of R Objects},
+    author = {Dirk Eddelbuettel with contributions by Antoine Lucas and Jarek Tuszynski and Henrik Bengtsson and Simon Urbanek and Mario Frasca and Bryan Lewis and Murray Stokely and Hannes Muehleisen and Duncan Murdoch and Jim Hester and Wush Wu and Qiang Kou and Thierry Onkelinx and Michel Lang and Viliam Simko and Kurt Hornik and Radford Neal and Kendon Bell and Matthew {de Queljoe} and Ion Suruceanu and Bill Denney and Dirk Schumacher and and Winston Chang.},
+    year = {2021},
+    note = {R package version 0.6.29},
+    url = {https://CRAN.R-project.org/package=digest},
+  }
+
+@misc{
+  Anonymizer2024, 
+  title={Anonymizer: Anonymize data containing personally identifiable information (PII) in R}, 
+  url={https://github.com/paulhendricks/anonymizer}, 
+  author={Hendricks, Paul}, 
+  note = {\url{https://github.com/paulhendricks/anonymizer} [Accessed: 29-08-2024]}
+}  
+  
+@misc{
+  deidentifyr2024, 
+  title={Deidentifyr}, 
+  url={https://github.com/wilkox/deidentifyr}, 
+  author={Wilkins, David}, 
+  note = {\url{https://github.com/wilkox/deidentifyr} [Accessed: 29-08-2024]}
+}  
+
+@inproceedings{wang2005break,
+  title={How to break MD5 and other hash functions},
+  author={Wang, Xiaoyun and Yu, Hongbo},
+  booktitle={Annual international conference on the theory and applications of cryptographic techniques},
+  pages={19--35},
+  year={2005},
+  organization={Springer}
+}
+
+@incollection{szikora2022end,
+  title={The end of encryption?--The era of quantum computers},
+  author={Szikora, P{\'e}ter and Laz{\'a}nyi, Korn{\'e}lia},
+  booktitle={Security-Related Advanced Technologies in Critical Infrastructure Protection: Theoretical and Practical Approach},
+  pages={61--72},
+  year={2022},
+  publisher={Springer}
+}
+
