### 2024-08-29 Stat-Cook/deident#5 @Stat-Cook
Section added to paper drawing comparisons.  NB: the two most focussed on deidentification (anonymizer and deidenifyr) are both now unpublished and feature restricted to salt+hash.
