### 2024-08-29 5c65bff
Add 'starwars' data for vignettes
