### 2024-08-29 4712383
Add extra vignettes for methods.
