+++ _pkgdown.yml
@@ -17,6 +17,7 @@ reference:
 - title: "Dataset"
   contents: 
   - ShiftsWorked
+  - starwars
 - title: "Utilities"
   contents: 
   - category_blur
