### 2024-08-29 #7157 @Stat-Cook
@PatrickRWright with regards to previous work done in R (https://github.com/Stat-Cook/deident/issues/5) the paper has been updated with a subsection discussing what has been done in R previously.  The two key packages listed by kamran for the removal of PID (anonymizer and deidentfyr) appear to both be out of support (last commit's 7 years and 5 years ago respectively) and neither is currently on CRAN (anonymizer was on CRAN but has been removed) or has been published elsewhere.
