### 2024-08-29 9c7abb5
Implement additional vignettes
