### 2024-08-30 #7157 @PatrickRWright
Please add a code coverage badge to the repo: https://github.com/Stat-Cook/deident/issues/7
