### 2024-08-30 #7157 @PatrickRWright
I strongly suggest considering a CRAN submission. Its overhead, yes, but in the end the quality of the work is improved.
https://github.com/Stat-Cook/deident/issues/8
