### 2024-08-30 Stat-Cook/deident#9 @PatrickRWright
Add "Community guidelines"
For the following points:
1) Contribute to the software
2) Report issues or problems with the software
3) Seek support
