### 2024-08-30 #7157 @PatrickRWright
Add community guidelines:
https://github.com/Stat-Cook/deident/issues/9
