### 2024-08-30 Stat-Cook/deident#7 @Stat-Cook
Afraid we have no familiarity with badges - can you recommend a method that coincides with `covr`?
