### 2024-08-30 #7157 @PatrickRWright
Unless I'm mistaken there is no automatic pipeline checking the package. Hence I suggest setting this up so that humans do not need to check what computers can do:
https://github.com/Stat-Cook/deident/issues/10
