### 2024-08-30 Stat-Cook/deident#7 @PatrickRWright
Have a look at the help here `?usethis::use_badge()`. Generally the `usethis` package will provide many boilerplate setups that will probably help you to "hit the ground running".

Its also possible to look at other popular packages to see how they set things up. e.g. https://github.com/tidyverse/tidyr
