### 2024-08-30 Stat-Cook/deident#11 @PatrickRWright
References in the paper for the basic theory of deidentification
Are there central references for the theory on methods to deidentify data independent of implementations?
