### 2024-08-30 #7157 @PatrickRWright
I assume the theoretical basis is described somewhere. It would be helpful to have references in the paper for the basic theory of deidentification:
https://github.com/Stat-Cook/deident/issues/11
