### 2024-08-30 Stat-Cook/deident#7 @Stat-Cook
Ahh brilliant - `usethis::use_test` is definitely an old friend.
