+++ tests/testthat/test-doc_templates.R
@@ -1,0 +1,6 @@
+test_that("doctemplates works", {
+
+  expect_equal(class(serialize.desc()), "character")
+  
+  expect_equal(class(transform.desc()), "character")
+})
