+++ tests/testthat/test-lookup_encrypter.R
@@ -1,0 +1,9 @@
+test_that("lookup encrypter", {
+  enc <- LookupEncrypter$new(hash_key = "hash", seed="seed")
+
+  enc$transform(starwars$name)
+  
+  expect_true(
+    length(enc$lookup) > 0
+  )
+})
