+++ tests/testthat/test-dot_deident.R
@@ -1,0 +1,17 @@
+test_that("BaseDotDeident works", {
+  bdd <- BaseDotDeident$new()
+  
+  .f <- function(...){
+    enquos(...)
+  }
+
+  quo <- .f(x, y, z)
+  
+  .serialize <- bdd$serialize(quo)
+  
+  expect_equal(
+    length(.serialize$args), 
+    3
+  )
+  
+})
