+++ man/add_shuffle.Rd
@@ -16,6 +16,7 @@ have for shuffling to be performed.  If the variable has length less than \code{
 values are repalced with \code{NA}s.}
 }
 \value{
+d
 A 'DeidentList' representing the untrained transformation pipeline.
 The object contains fields:
 \itemize{
