+++ man/add_mutate.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/api.R
+\name{add_mutate}
+\alias{add_mutate}
+\title{Add a 'tidyverse' mutate to the pipeline.}
+\usage{
+add_mutate(object, ...)
+}
+\description{
+Add a 'tidyverse' mutate to the pipeline.
+}
