+++ man/add_tidy.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/api.R
+\name{add_tidy}
+\alias{add_tidy}
+\title{Add a 'tidyverse' function to the pipeline.}
+\usage{
+add_tidy(object, ..., fn = function(x) x)
+}
+\description{
+Add a 'tidyverse' function to the pipeline.
+}
