+++ DESCRIPTION
@@ -23,22 +23,23 @@ Suggests:
     testthat (>= 3.0.0),
     checkmate,
     knitr,
-    rmarkdown
+    rmarkdown,
+    roxygen2
 Imports:
-  R6,
-  dplyr,
-  openssl,
-  tidyselect,
-  rlang,
-  glue,
-  purrr,
-  stringr,
-  yaml,
-  readr,
-  openxlsx,
-  lemon,
-  withr,
-  fs
+    R6,
+    dplyr,
+    openssl,
+    tidyselect,
+    rlang (>= 0.4.11),
+    glue,
+    purrr,
+    stringr,
+    yaml,
+    readr,
+    openxlsx,
+    lemon,
+    withr,
+    fs
 Depends: 
     R (>= 2.10)
 VignetteBuilder: knitr
