+++ NAMESPACE
@@ -4,10 +4,8 @@ S3method(apply_deident,character)
 S3method(apply_deident,data.frame)
 S3method(apply_to_data_frame,BaseDeident)
 S3method(apply_to_data_frame,DeidentTask)
-S3method(apply_to_data_frame,R6)
 S3method(apply_to_data_frame,R6ClassGenerator)
 S3method(apply_to_data_frame,character)
-S3method(apply_to_data_frame,list)
 S3method(create_deident,BaseDeident)
 S3method(create_deident,GrouperR6ClassGenerator)
 S3method(create_deident,R6ClassGenerator)
@@ -25,6 +23,8 @@ S3method(serialize,BaseDeident)
 S3method(serialize,DeidentList)
 S3method(serialize,DeidentTask)
 S3method(str,DeidentTask)
+export(":=")
+export(.data)
 export(BaseDeident)
 export(Blurer)
 export(Drop)
@@ -45,10 +45,14 @@ export(add_shuffle)
 export(add_ungroup)
 export(apply_deident)
 export(apply_to_data_frame)
+export(as_label)
+export(as_name)
 export(category_blur)
 export(create_deident)
 export(deident)
 export(deident_job_from_folder)
+export(enquo)
+export(enquos)
 export(from_yaml)
 export(lognorm_noise)
 export(new_deident)
@@ -63,21 +67,23 @@ importFrom(dplyr,select)
 importFrom(glue,glue)
 importFrom(openssl,sha256)
 importFrom(openxlsx,read.xlsx)
+importFrom(purrr,as_mapper)
 importFrom(purrr,imap)
 importFrom(purrr,map)
 importFrom(purrr,reduce)
 importFrom(purrr,simplify)
 importFrom(readr,read_csv)
 importFrom(readr,read_tsv)
+importFrom(rlang,":=")
+importFrom(rlang,.data)
 importFrom(rlang,as_label)
+importFrom(rlang,as_name)
 importFrom(rlang,enquo)
 importFrom(rlang,enquos)
-importFrom(rlang,sym)
 importFrom(stats,rlnorm)
 importFrom(stats,rnorm)
 importFrom(stats,sd)
 importFrom(stringr,str_detect)
 importFrom(stringr,str_remove)
-importFrom(tidyselect,eval_select)
 importFrom(yaml,read_yaml)
 importFrom(yaml,write_yaml)
