### 2024-08-30 6aed75e
Add code coverage badge
