+++ README.md
@@ -1,5 +1,7 @@
 # deident
 
+[![codecov](https://codecov.io/gh/Stat-Cook/deident/graph/badge.svg?token=1B5QA2JIP6)](https://codecov.io/gh/Stat-Cook/deident)
+
 Tools for creating and applying a semi-automated persistent de-identification 
 pipeline to sensitive data.  The package is implemented in R. 
 
