### 2024-08-30 11294b6
Add cmd check + badge
