+++ README.md
@@ -1,6 +1,9 @@
 # deident
+  <!-- badges: start -->
+  [![R-CMD-check](https://github.com/Stat-Cook/deident/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/Stat-Cook/deident/actions/workflows/R-CMD-check.yaml)
+  [![codecov](https://codecov.io/gh/Stat-Cook/deident/graph/badge.svg?token=1B5QA2JIP6)](https://codecov.io/gh/Stat-Cook/deident)
+  <!-- badges: end -->
 
-[![codecov](https://codecov.io/gh/Stat-Cook/deident/graph/badge.svg?token=1B5QA2JIP6)](https://codecov.io/gh/Stat-Cook/deident)
 
 Tools for creating and applying a semi-automated persistent de-identification 
 pipeline to sensitive data.  The package is implemented in R. 
