+++ man/new_deident.Rd
@@ -12,7 +12,6 @@ new_deident(object, ..., encrypter)
 \item{...}{variables to be transformed.}
 }
 \value{
-d
 A 'DeidentList' representing the untrained transformation pipeline.
 The object contains fields:
 \itemize{
