### 2024-08-30 Stat-Cook/deident#10 @Stat-Cook
CMD Checks added and button added to ReadMe
