### 2024-08-30 Stat-Cook/deident#7 @Stat-Cook
code coverage badge added - currently at 94%
