### 2024-08-30 86117c6
Add `Getting help` to cover community notice.
