+++ README.md
@@ -17,3 +17,10 @@ library("deident")
 
 The documentation for `deident` is hosted via github pages [here](https://stat-cook.github.io/deident/).  We suggest looking at the [`Worked Example`](https://stat-cook.github.io/deident/articles/worked_example.html) vignette to understand the intended workflow.
 
+# Getting help
+
+If you encounter a bug, please file a minimal reproducible example on [github](https://github.com/Stat-Cook/deident/issues).  Any feature requests, or requests for guidance on using the package can also be submitted this way though please [read the documentation first](https://stat-cook.github.io/deident/).  
+
+Contribution to `deident` is welcome, to do so please open an [issue](https://github.com/Stat-Cook/deident/issues) so the development team can check it is compatible with the design principles or feel free to fork the repository.
+
+
