+++ paper.bib
@@ -120,3 +120,49 @@ urldate = {2023-07-08}
   publisher={Springer}
 }
 
+@book{szikora2022end,
+  title={The end of encryption?--The era of quantum computers},
+  author={Szikora, P{\'e}ter and Laz{\'a}nyi, Korn{\'e}lia},
+  booktitle={Security-Related Advanced Technologies in Critical Infrastructure Protection: Theoretical and Practical Approach},
+  pages={61--72},
+  year={2022},
+  publisher={Springer}
+}
+
+@booklet{ICOAnonymisation,
+  title        = {Anonymisation: managing data protection risk code of practice},
+  author       = {ICO},
+  year         = 2012,
+  month        = {November},
+  address      = {Wycliffe House, Cheshire}
+}
+
+@book{NIST2015,
+  title        = {De-Identification of Personal Information},
+  author       = {Garfinkel, Simson L.},
+  publisher = {NIST},
+  year         = 2015,
+  month        = {October},
+  address      = {Gaithersburg, MD}
+}
+
+
+@misc{
+  integrateIO, 
+  title={6 Steps for Data Pseudonymization}, 
+  url={https://www.integrate.io/blog/6-steps-to-pseudonymize-pii/#steps}, 
+  author={Integrate.IO}, 
+  note = {\url{https://www.integrate.io/blog/6-steps-to-pseudonymize-pii/#steps} [Accessed: 02-09-2024]}
+}  
+
+@Article{
+  tidyverse,
+     title = {Welcome to the {tidyverse}},
+     author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
+     year = {2019},
+     journal = {Journal of Open Source Software},
+     volume = {4},
+     number = {43},
+     pages = {1686},
+     doi = {10.21105/joss.01686},
+}
