### 2024-09-03 21a20ea
Update paper intro
