+++ paper.Rmd
@@ -7,12 +7,10 @@ bibliography: paper.bib
 
 # Summary
 
-Research in the discipline of health data is of increasing interest due to the 
-perception that artificial intelligence (AI) and machine learning (ML) 
-techniques have the promise to bridge gaps in the large, resource-limited 
-sector [@harwich2018thinking; @wilson2019high; @nelson2019predicting; @yu2018artificial].  One key aspect of working within health data is the necessity to work  within the ethical and legal frameworks of research with human participants[@sales2000ethics; @UKRI], notably around the risks posed by the processing of personally identifiable data (PID) and pseudo-PID (variables which if used together can identify an individual)[@ICO].  
+The delivery of quality health care is a constant act of balancing  demand  against  capacity,  with emerging, data intensive, artificial intelligence (AI) and machine learning (ML) approaches poised to bridge gaps in the large, resource-limited 
+sector [@harwich2018thinking; @wilson2019high; @nelson2019predicting; @yu2018artificial].  The scale of data required in such projects magnifies the importance of  existing  ethical and legal frameworks for research with human participants[@sales2000ethics; @UKRI], notably around the risks posed by the processing of personally identifiable data (PID) and pseudo-PID (variables which if used together can identify an individual)[@ICO].  
 
-One approach to dealing with PID concerns is to apply transformations to the data, e.g. encryption of names, or aggregation of ages, which  can limit the risk of identification  at the cost of nuance[@tachepun2020data].   Hence, we demonstrate an extendable package of tools for the implementation  and application of deidentification techniques to panel datasets.
+One approach to dealing with PID concerns is to apply transformations to the data, e.g. encryption of names, or aggregation of ages, which  can limit the risk of identification  at the cost of nuance[@tachepun2020data].   Hence, we demonstrate an extendable package of tools for the implementation  and application of deidentification techniques to panel data sets.
 
 # Statement of need
 
