### 2024-09-05 #7078 @spholmes
Hi @nrennie, I wonder if you are interested in helping us review this submission to JOSS?
