### 2024-09-05 Stat-Cook/deident#11 @Stat-Cook
Key references added:
* ICO guidenlines on anonymization
* Garfinkel 2015
* Integrate.IO (grey literature)
