### 2024-09-05 #7078 @PeerHerholz
> Hello @PeerHerholz As an author of a paper on data in JOSS, we thought you might be able to help us out with a review of this submission on the R package Deident. Would you be able to help us out here? Thanks Susan

Hi @spholmes and everyone,

I totally missed this, I'm very sorry.
It sounds super interesting but I'm not that familiar with `R` and especially the development/publishing aspects. Thus, I can't be that helpful here. I hope you'll find a fitting reviewer soon.

Sorry again.

Best, Peer
