### 2024-09-16 #7078 @nrennie
> Hi @nrennie, I wonder if you are interested in helping us review this submission to JOSS?

Yes, I'd be happy to review it
