### 2024-09-16 #7078 @spholmes
@nrennie : Many Thanks
You'll see a new review issue here : https://github.com/openjournals/joss-reviews/issues/7157
please generate your checklist there with the editorialbot by asking it:
`@editorialbot generate my checklist`
over there on the review issue.
