### 2024-10-22 #7157 @Stat-Cook
> "Contribution and authorship": I can't judge the contribution of the authors aside from @Stat-Cook since they do not appear in the commit history of the software. I suggest stating individual contributions in a dedicated section of the paper.

Paper now has a dedicated 'Contributions' section.
