### 2024-10-22 #7157 @Stat-Cook
> It would be nice to have vignette-stlye examples fir all mentioned de-identification methods.

Extra vigenttes have been added (https://github.com/Stat-Cook/deident/issues/6) to cover all methods.
