+++ vignettes/worked_example.Rmd
@@ -48,7 +48,7 @@ knitr::opts_chunk$set(
 library(deident)
 ```
 
-To assist new users, this vignette serves as a worked example of building, applying and saving a de-identification pipeline.  Due to the risk of suppling data with genuine PID to be redacted, we instead supply a synthetic data set comprising of artificial PID, `ShiftsWorked`.
+To assist new users, this vignette serves as a worked example of building, applying and saving a de-identification pipeline.  Due to the risk of supplying data with genuine PID to be redacted, we instead supply a synthetic data set comprising of artificial PID, `ShiftsWorked`.
 
 The data set consists of 3,100 observations where each row lists the work done (shift status, actual work done, and daily pay) by an individual on a given day.  `ShiftsWorked` consists of 7 variables:
 
