### 2024-10-23 Stat-Cook/deident#8 @Stat-Cook
Began process of CRAN submission.  Checks pass on R-devel distro.
