### 2024-10-28 6661040
Auto stash before merge of "master" and "origin/master"
