+++ R/test_imports.R
@@ -1,0 +1,6 @@
+#' @importFrom fs file_delete
+#' @importFrom withr defer
+#' 
+#' @importFrom lemon lemon_print
+#' 
+f <- function(){}
