+++ NEWS.md
@@ -1,3 +1,4 @@
-# deident (development version)
+# deident 1.0
   
 * Implemented a suite of methods for PID removal.
+
