+++ NAMESPACE
@@ -64,26 +64,34 @@ importFrom(dplyr,group_by)
 importFrom(dplyr,group_modify)
 importFrom(dplyr,mutate)
 importFrom(dplyr,select)
+importFrom(fs,file_delete)
 importFrom(glue,glue)
+importFrom(lemon,lemon_print)
 importFrom(openssl,sha256)
 importFrom(openxlsx,read.xlsx)
 importFrom(purrr,as_mapper)
 importFrom(purrr,imap)
 importFrom(purrr,map)
 importFrom(purrr,reduce)
+importFrom(purrr,reduce2)
 importFrom(purrr,simplify)
 importFrom(readr,read_csv)
 importFrom(readr,read_tsv)
 importFrom(rlang,":=")
 importFrom(rlang,.data)
+importFrom(rlang,abort)
 importFrom(rlang,as_label)
 importFrom(rlang,as_name)
 importFrom(rlang,enquo)
 importFrom(rlang,enquos)
+importFrom(rlang,sym)
 importFrom(stats,rlnorm)
 importFrom(stats,rnorm)
 importFrom(stats,sd)
 importFrom(stringr,str_detect)
 importFrom(stringr,str_remove)
+importFrom(tidyselect,eval_select)
+importFrom(utils,head)
+importFrom(withr,defer)
 importFrom(yaml,read_yaml)
 importFrom(yaml,write_yaml)
