### 2024-10-28 #7157 @Stat-Cook
@PatrickRWright  I think we've been able to address all the comments with the caveat that the CRAN submission has been started but not finalized.  If you could review and let us know the next steps we'd be grateful!
