### 2024-10-30 Stat-Cook/deident#7 @PatrickRWright
@Stat-Cook looks like the badge is broken.
