### 2024-10-30 #7157 @PatrickRWright
Basically everything has been addressed. Thanks!

I'd like to wait with a final functional test until the package is accepted and up on CRAN.
