### 2024-11-13 184fc89
Add encryption definition
