+++ paper.Rmd
@@ -26,6 +26,7 @@ This package was designed considering the ICO guidelines on anonymization [@ICOA
 *	shuffling – replacement of columns by a random sample without replacement
 *	bluring – the aggregation of numeric or categorical data according to specified rules
 *	perturbation – the addition of user-defined random noise to a numeric variable
+* encryption - the consistent replacement of a string by an alpha-numeric hash using an encryption key and salt
 
 Following the design principles of the existing tidyverse[@tidyverse] domain for ease of
 adoption.  The package includes tools to create a single pipeline for the 
@@ -35,7 +36,7 @@ the pipeline to/ via yaml.  This approach allows a researcher to design and
 implement an appropriate de-identification plan and deliver it to the research 
 support/ business intelligence team of an organisation with limited knowledge 
 of the sensitive data.  The supply of an easy method for de-identifying data
-sets which require little scripting knowledge by Trust staff may aid in 
+sets which requires little scripting knowledge by Trust staff may aid in 
 overcoming several information governance risks that keep operational data 
 siloed within health Trusts.  
 
