+++ paper.Rmd
@@ -22,11 +22,11 @@ This implementation of the "deident" methods is in R, chosen due to the increase
 This package was designed considering the ICO guidelines on anonymization [@ICOAnonymisation] and draws on pre-existing methodologies and naming conventions[@NIST2015; @integrateIO]. The package implemented de-identification via:   
 
 *	pseudonymization – the consistent replacement of a string by a random string
-*	encryption 
+* encryption - the consistent replacement of a string by an alpha-numeric hash using an encryption key and salt
 *	shuffling – replacement of columns by a random sample without replacement
 *	bluring – the aggregation of numeric or categorical data according to specified rules
 *	perturbation – the addition of user-defined random noise to a numeric variable
-* encryption - the consistent replacement of a string by an alpha-numeric hash using an encryption key and salt
+
 
 Following the design principles of the existing tidyverse[@tidyverse] domain for ease of
 adoption.  The package includes tools to create a single pipeline for the 
