### 2024-11-13 #7157 @Stat-Cook
Update - thanks for the last few notes.  We've now made a submission to CRAN.  There was a slight delay (codecov badge throws an error on submission to CRAN, and 'fixing' it for CRAN breaks the badge) which we've fixed by excluding the README from the build.

Will let you know when it's accepted on CRAN.
