### 2024-11-20 #7157 @Stat-Cook
@PatrickRWright  package has been accepted to CRAN https://cran.r-project.org/web/packages/deident/index.html
