### 2024-11-22 #7157 @PatrickRWright
Ok, from my side its ok now. Thank you for this library.

I would like to stress that I did not perform a review of the source code. My core focus was on adherence to good R Package development guidelines and checking of the documentation and the paper text. Hence I reviewed this as a potential user not as a subject matter expert for the methods being implemented. Having said that, from going over the vignettes it looks fine to me.

I'm also glad you followed through on the CRAN submission. Thanks for going the extra mile.

Minor point, which does not need to be addressed for this review but is generally useful for package development: I suggest looking into code linting. Pick/Establish a style guide and lint your code according to the rules. My suggestion would be to stick with the tidyverse style guide since its already there and a well accepted standard.
