### 2024-11-22 #7157 @Stat-Cook
Thanks for the review @PatrickRWright - we've been looking at lintr for just that thought.  Probably will be rolled out if the code base grows.
