### 2024-11-26 Stat-Cook/deident#14 @nrennie
Choice of function name 
Why the non-standard spelling of "Blurer"? Repeatedly switch between "Blurring" and "Bluring". The spelling of the function name doesn't match the way it would be spelled in a sentence which is confusing. The same spelling mistake appears in the paper. I suggest renaming the function, or creating an alias.

If this is a deliberate design decision, maybe a note somewhere?
