### 2024-11-26 #7157 @nrennie
Choice of function name: https://github.com/Stat-Cook/deident/issues/14
