### 2024-11-26 #7157 @nrennie
Vignettes could be tidier and clearer: https://github.com/Stat-Cook/deident/issues/15
