### 2024-11-26 #7157 @nrennie
Minor issues with paper: https://github.com/Stat-Cook/deident/issues/16
