### 2024-11-26 #7157 @nrennie
Initial review finished - they are mainly all minor issues with code documentation or paper formatting. The choice of function name for `Blurer()` is the only sort of major issue I see.
