### 2024-11-28 Stat-Cook/deident#15 @Stat-Cook
To track:

* ~~I would suggest re-ordering the vignettes to be in a more logical order e.g the "rationale" vignette should be at the top rather than in the middle of transformation examples, the the more advanced example "re-using" should be after the simple example.~~
* ~~In Transformations vignette, capitalise title~~
* ~~In "Options" section of transformation vignette, the code is commented out. Why? If you don't want the code to execute set eval to FALSE.~~
* In Transformations vignette, you say "To demonstrate the different transformation we supply a toy data set, df, comprising 26 " but none of the output is shown, so you don't actually show the transformation.
