### 2024-11-28 Stat-Cook/deident#15 @Stat-Cook
@nrennie quick check - for ordering of vignettes.  Is this just ordering in the pkgdown documentation?
