### 2024-11-28 Stat-Cook/deident#16 @Stat-Cook
Tracking:

* ~~In the babynames example in the paper, why do you convert it to a data frame? It already is a tibble and therefore a dataframe. This needs fixed or explained. I would suggest loading the deident package after setting up the example data. It doesn't need to be loaded in the babynames filtering code block.~~ Thanks for catching this, parse to data frame has been removed.  I think an early draft we created a data set to test on but replaced with existing data.
* ~~In software paper, double check formatting of when you are using single/double quotations, tick marks, and italics. Sometimes you put package names in code format using back ticks, other times in quotations. Be consistent.~~
* ~~There are a few references with n.d. but these are all on GitHub and the date can be found on GitHub.~~ We were unsure if the year should be date of last commit - updated to last commit date for now.
* ~~Some of the spacing in the paper is odd e.g. make sure space before a (.~~
* ~~You use the babynames data for examples in the paper, suggest including it in citation list.~~
* ~~Suggest using {pak} or {remotes} in GitHub installation instructions. {devtools} hasn't been the recommended method for a while~~ Changed to pak.
