### 2024-11-28 fae072c
Reorder vignettes in pkgdown
