+++ paper.bib
@@ -91,14 +91,16 @@ urldate = {2023-07-08}
   title={Anonymizer: Anonymize data containing personally identifiable information (PII) in R}, 
   url={https://github.com/paulhendricks/anonymizer}, 
   author={Hendricks, Paul}, 
+  year = {2017},
   note = {\url{https://github.com/paulhendricks/anonymizer} [Accessed: 29-08-2024]}
-}  
-  
+}
+
 @misc{
   deidentifyr2024, 
   title={Deidentifyr}, 
   url={https://github.com/wilkox/deidentifyr}, 
   author={Wilkins, David}, 
+  year = {2019},
   note = {\url{https://github.com/wilkox/deidentifyr} [Accessed: 29-08-2024]}
 }  
 
