### 2024-11-28 fdae2d0
Swap to `pak` repo for install.
