+++ paper.Rmd
@@ -49,7 +49,8 @@ Several packages have undergone development for the implementation of `encryptio
 To install the current version of the `deident` package, run the command
 
 ``` {r, eval=F}
-devtools::install_github("Stat-Cook/deident")
+# install.packages("pak")
+pak::pkg_install("Stat-Cook/deident")
 ```
 
 The core functionality of the package is the `deident` function. To demonstrate functionality we use a subset of the `babynames` data set consisting of the final two years.  
