+++ paper.md
@@ -29,14 +29,15 @@ intelligence (AI) and machine learning (ML) approaches poised to bridge
 gaps in the large, resource-limited sector (Harwich and Laycock 2018;
 Wilson 2019; Nelson et al. 2019; Yu, Beam, and Kohane 2018). The scale
 of data required in such projects magnifies the importance of existing
-ethical and legal frameworks for research with human participants(Sales
+ethical and legal frameworks for research with human participants (Sales
 and Folkman 2000; UKRI 2022), notably around the risks posed by the
 processing of personally identifiable data (PID) and pseudo-PID
-(variables which if used together can identify an individual)(ICO 2023).
+(variables which if used together can identify an individual) (ICO
+2023).
 
 One approach to dealing with PID concerns is to apply transformations to
 the data, e.g. encryption of names, or aggregation of ages, which can
-limit the risk of identification at the cost of nuance(Tachepun and
+limit the risk of identification at the cost of nuance (Tachepun and
 Thammaboosadee 2020). Hence, we demonstrate an extendable package of
 tools for the implementation and application of deidentification
 techniques to panel data sets.
@@ -63,7 +64,7 @@ the transforms can be ported to other languages.
 
 This package was designed considering the ICO guidelines on
 anonymization (ICO 2012) and draws on pre-existing methodologies and
-naming conventions(Garfinkel 2015; Integrate.IO, n.d.). The package
+naming conventions (Garfinkel 2015; Integrate.IO, n.d.). The package
 implemented de-identification via:
 
 -   pseudonymization – the consistent replacement of a string by a
@@ -77,33 +78,33 @@ implemented de-identification via:
 -   perturbation – the addition of user-defined random noise to a
     numeric variable
 
-Following the design principles of the existing tidyverse(Wickham et al.
-2019) domain for ease of adoption. The package includes tools to create
-a single pipeline for the application of a multi-step deidentification
-pipeline to multiple files stored within the same repository, alongside
-the option to serialize/ define the pipeline to/ via yaml. This approach
-allows a researcher to design and implement an appropriate
-de-identification plan and deliver it to the research support/ business
-intelligence team of an organisation with limited knowledge of the
-sensitive data. The supply of an easy method for de-identifying data
-sets which requires little scripting knowledge by Trust staff may aid in
-overcoming several information governance risks that keep operational
-data siloed within health Trusts.
+Following the design principles of the existing tidyverse (Wickham et
+al. 2019) domain for ease of adoption. The package includes tools to
+create a single pipeline for the application of a multi-step
+deidentification pipeline to multiple files stored within the same
+repository, alongside the option to serialize/ define the pipeline to/
+via yaml. This approach allows a researcher to design and implement an
+appropriate de-identification plan and deliver it to the research
+support/ business intelligence team of an organisation with limited
+knowledge of the sensitive data. The supply of an easy method for
+de-identifying data sets which requires little scripting knowledge by
+Trust staff may aid in overcoming several information governance risks
+that keep operational data siloed within health Trusts.
 
 # Comparison to existing R packages
 
 Several packages have undergone development for the implementation of
 `encryption` methods to minimize identifiability within data sets
-(e.g. ‘anonymizer’(Hendricks 2017), ‘deidentifyr’(Wilkins 2019) and
-‘digest’(Antoine Lucas et al. 2021)). While these packages implement a
-variety of encryption tools, such systems are not infallible(Wang and Yu
-2005; Szikora and Lazányi 2022 ) especially if the data being encrypted
-is drawn from a known domain such as common names. This package
-introduces the stateful ‘pseudonymization’ method by which string
[…]
