+++ paper.Rmd
@@ -8,18 +8,18 @@ bibliography: paper.bib
 # Summary
 
 The delivery of quality health care is a constant act of balancing  demand  against  capacity,  with emerging, data intensive, artificial intelligence (AI) and machine learning (ML) approaches poised to bridge gaps in the large, resource-limited 
-sector [@harwich2018thinking; @wilson2019high; @nelson2019predicting; @yu2018artificial].  The scale of data required in such projects magnifies the importance of  existing  ethical and legal frameworks for research with human participants[@sales2000ethics; @UKRI], notably around the risks posed by the processing of personally identifiable data (PID) and pseudo-PID (variables which if used together can identify an individual)[@ICO].  
+sector  [@harwich2018thinking; @wilson2019high; @nelson2019predicting; @yu2018artificial].  The scale of data required in such projects magnifies the importance of  existing  ethical and legal frameworks for research with human participants  [@sales2000ethics; @UKRI], notably around the risks posed by the processing of personally identifiable data (PID) and pseudo-PID (variables which if used together can identify an individual)  [@ICO].  
 
-One approach to dealing with PID concerns is to apply transformations to the data, e.g. encryption of names, or aggregation of ages, which  can limit the risk of identification  at the cost of nuance[@tachepun2020data].   Hence, we demonstrate an extendable package of tools for the implementation  and application of deidentification techniques to panel data sets.
+One approach to dealing with PID concerns is to apply transformations to the data, e.g. encryption of names, or aggregation of ages, which  can limit the risk of identification  at the cost of nuance  [@tachepun2020data].   Hence, we demonstrate an extendable package of tools for the implementation  and application of deidentification techniques to panel data sets.
 
 # Statement of need
 
-In order to broaden the access to sensitive data, data handlers need an auditable open process for the application of data transforms that limit the risk of personal identification (sometimes refereed to as 'data masking' transformations).  Several production scale systems exist for applying data masking transformations (e.g. [Delphix](https://www.delphix.com/), [K2View](https://www.k2view.com), and [Accutive](https://accutive.com/)), but these are often expensive and cumbersome, limiting uptake in the domain of research.  The research community hence requires a simple toolbox for the application of several common 'data masking transforms' implemented  in an open source language.
+In order to broaden the access to sensitive data, data handlers need an auditable open process for the application of data transforms that limit the risk of personal identification (sometimes refereed to as 'data masking' transformations).  Several production scale systems exist for applying data masking transformations (e.g.  [Delphix](https://www.delphix.com/),  [K2View](https://www.k2view.com), and  [Accutive](https://accutive.com/)), but these are often expensive and cumbersome, limiting uptake in the domain of research.  The research community hence requires a simple toolbox for the application of several common 'data masking transforms' implemented  in an open source language.
 
 This implementation of the "deident" methods is in R, chosen due to the increased adoption of open source software into the working practices of the NHS (driven by the growth on the NHS-R community and similar groups).  However, the underlying specification, and implementation of the transforms can be ported to other languages.
 
 
-This package was designed considering the ICO guidelines on anonymization [@ICOAnonymisation] and draws on pre-existing methodologies and naming conventions[@NIST2015; @integrateIO]. The package implemented de-identification via:   
+This package was designed considering the ICO guidelines on anonymization  [@ICOAnonymisation] and draws on pre-existing methodologies and naming conventions  [@NIST2015; @integrateIO]. The package implemented de-identification via:   
 
 *	pseudonymization – the consistent replacement of a string by a random string
 * encryption - the consistent replacement of a string by an alpha-numeric hash using an encryption key and salt
[…]
