### 2024-12-01 Stat-Cook/deident#15 @nrennie
@Stat-Cook Yes, doing that should also re-order them on the CRAN page: https://cran.r-project.org/web/packages/deident/index.html
