+++ pkgdown.yml
@@ -1,0 +1,14 @@
+pandoc: 3.1.11
+pkgdown: 2.1.1
+pkgdown_sha: ~
+articles:
+  blur_example: blur_example.html
+  encrypt_example: encrypt_example.html
+  numeric_blur_example: numeric_blur_example.html
+  perturb_example: perturb_example.html
+  rationale: rationale.html
+  reusing_methods: reusing_methods.html
+  shuffle_example: shuffle_example.html
+  transformations: transformations.html
+  worked_example: worked_example.html
+last_built: 2024-12-02T14:52Z
