+++ vignettes/shuffle_example.Rmd
@@ -29,18 +29,19 @@ head(starwars)
 Inspection of the data shows `species` can be a unique identifier (e.g. 'Admiral Ackbar' is the only 'Mon Calamari') so we may consider aggregating species:
 
 ``` {r}
-starwars |> 
+starwars |>
   dplyr::filter(species == "Mon Calamari")
 ```
 
 However, while knowing someone is 'Human' does not have the same effect, if we also knew they were from 'Coruscant' and had 'blond' hair (each of which is not uniquely identifiable) if used in combination we reduce the data to a single case: 
 
 ``` {r}
-starwars |> 
-  dplyr::filter(species == "Human", 
-                homeworld == "Coruscant",
-                hair_color == "blond"
-                )
+starwars |>
+  dplyr::filter(
+    species == "Human",
+    homeworld == "Coruscant",
+    hair_color == "blond"
+  )
 ```
 
 Hence, individual columns can contain useful information but we may not wish to disclose the inter-variable correlations.  To aid with this, we introduce the `shuffling` method  which performs column wise sampling without replacement:
@@ -50,7 +51,7 @@ NB: we set a random seed using `set.seed` here for reproducibility.  We recommen
 ``` {r}
 set.seed(101)
 
-shuffle_pipe <- starwars |> 
+shuffle_pipe <- starwars |>
   add_shuffle(species, homeworld, hair_color)
 
 new_starwars <- apply_deident(starwars, shuffle_pipe)
@@ -61,10 +62,11 @@ head(new_starwars)
 A `Shuffle` hence preserves the column summaries, e.g. modal values and distributions, but breaks inter-column behaviours which might lead to identification.
 
 ``` {r}
-new_starwars |> 
-  dplyr::filter(species == "Human", 
-                homeworld == "Coruscant"
-                )
+new_starwars |>
+  dplyr::filter(
+    species == "Human",
+    homeworld == "Coruscant"
+  )
 ```
 
 ### `Grouped Shuffling`
@@ -72,8 +74,8 @@ new_starwars |>
 Clearly there will be situations in which inter-variable dependencies are key to our understanding of the data, and we may wish to preserve the column metrics within strata.  Such a situation is foreseen, and 'shuffling' can be performed within a grouped data set as easily as on the whole data:
 
 ``` {r}
-grouped_shuffle_pipe <- starwars |> 
-  add_group(gender) |> 
+grouped_shuffle_pipe <- starwars |>
+  add_group(gender) |>
   add_shuffle(species, homeworld, hair_color) |>
   add_ungroup()
 
