+++ vignettes/blur_example.Rmd
@@ -28,7 +28,7 @@ A simple 'blur' might be to change the taxonomy of 'Shift' e.g. combine 'Day' an
 ``` {r}
 shift_blur <- c("Day" = "Working", "Night" = "Working")
 blur_pipe <- ShiftsWorked |>
-  add_blur(Shift, blur=shift_blur)
+  add_blur(Shift, blur = shift_blur)
 
 apply_deident(ShiftsWorked, blur_pipe)
 ```
@@ -62,8 +62,8 @@ And the vector returned can be passed into a new pipeline as before.
 
 ``` {r}
 species_pipe <- starwars |>
-  add_blur(species, blur=human_blur)
-  
+  add_blur(species, blur = human_blur)
+
 new_starwars <- apply_deident(starwars, species_pipe)
 
 table(new_starwars$species)
