+++ R/create_deident.R
@@ -1,18 +1,15 @@
-#' Create a deident pipeline 
-#' 
+#' Create a deident pipeline
+#'
 #' @param method A deidentifier to initialize.
-#' @param ... list of variables to be deidentified. NB: key word arguments will 
+#' @param ... list of variables to be deidentified. NB: key word arguments will
 #' be passed to method at initialization.
-#' 
+#'
 #' @export
-create_deident <- function(method, ...){
-
-
-  UseMethod("create_deident",method)
+create_deident <- function(method, ...) {
+  UseMethod("create_deident", method)
 }
 
-create_deident.BaseDeident <- function(method, ...){
-
+create_deident.BaseDeident <- function(method, ...) {
   #' @exportS3Method
   #' @importFrom rlang enquos
   #'
@@ -28,24 +25,23 @@ create_deident.BaseDeident <- function(method, ...){
   l
 }
 
-create_deident.R6ClassGenerator <- function(method, ...){
+create_deident.R6ClassGenerator <- function(method, ...) {
   #' @exportS3Method
   #' @export
 
   dots <- arg_kwarg(...)
 
   init.list <- purrr::map(dots$kwargs, rlang::quo_get_expr)
-    
-  .tra <- do.call(method$new, init.list)
-  
-  create_deident(.tra, !!!dots$args)
 
+  .tra <- do.call(method$new, init.list)
+
+  create_deident(.tra, !!!dots$args)
 }
 
-create_deident.GrouperR6ClassGenerator <- function(method, ...){
+create_deident.GrouperR6ClassGenerator <- function(method, ...) {
   #' @exportS3Method
   #' @export
-  #' 
+  #'
 
   l <- list(
     variables = c(),
@@ -56,31 +52,30 @@ create_deident.GrouperR6ClassGenerator <- function(method, ...){
   l
 }
 
-create_deident.list <- function(method, ...){
+create_deident.list <- function(method, ...) {
   #' @exportS3Method
-  
+
   create_deident(method$Type, !!!method$variables, !!!method$args)
 }
 
-create_deident.character <- function(method, ...){
+create_deident.character <- function(method, ...) {
   #' @exportS3Method
   #' @export
 
   implemented_transforms <- implemented_transforms_base.f()
   .tra <- implemented_transforms[[method]]
-  
-  if (is.null(.tra)){
+
+  if (is.null(.tra)) {
     abort(glue::glue("Transform method {method} not implemented"))
   }
-  
+
   create_deident(.tra, ...)
 }
 
 
-str.DeidentTask <- function(object, ...){
+str.DeidentTask <- function(object, ...) {
   #' @exportS3Method
   labels <- quosure_as_labels(object$variables)
   glue("'{object$method$str()}' on variable(s) \\
        {paste(labels, collapse=', ')}")
 }
-
