+++ tests/testthat/test-user_defined_deident.R
@@ -1,21 +1,21 @@
 test_that("UserDefinedDeident works", {
-  method <- function(vec){
-     rep("", length(vec))
+  method <- function(vec) {
+    rep("", length(vec))
   }
-  
+
   udd <- UserDefinedDeident$new(method)
   .vec <- udd$transform(letters)
-  
+
   expect_equal(
-    .vec, 
+    .vec,
     rep("", 26)
   )
-  
+
   udd2 <- deident_from_func(method)
   .vec2 <- udd2$transform(letters)
-  
+
   expect_equal(
-    .vec2, 
+    .vec2,
     rep("", 26)
   )
 })
