+++ tests/testthat/test-pseudonomizer.R
@@ -1,11 +1,11 @@
-expect_character <- function(object){
+expect_character <- function(object) {
   act <- quasi_label(rlang::enquo(object), arg = "object")
 
   testthat::expect(
     is.character(act$val),
     sprintf("%s is of type %s, not 'character'", act$lab, act$val)
   )
-  
+
   invisible(act$val)
 }
 
@@ -15,18 +15,16 @@ test_that("Psudonomizer works", {
     A = sample(LETTERS[1:4], n, T),
     B = rnorm(n)
   )
-  
+
   .tra <- Pseudonymizer$new()
   .tra$transform(df$A)
   df$A
   expect_character(.tra$transform(df$A))
   expect_equal(length(.tra$transform(df$A)), n)
-  
-  expect_warning(.tra$transform(df$B, parse_numerics=F))
+
+  expect_warning(.tra$transform(df$B, parse_numerics = F))
 
   .mut <- .tra$mutate(df, A)
   expect_data_frame(.mut)
   expect_equal(nrow(.mut), n)
-
 })
-
