+++ tests/testthat/test-dot_deident.R
@@ -1,17 +1,16 @@
 test_that("BaseDotDeident works", {
   bdd <- BaseDotDeident$new()
-  
-  .f <- function(...){
+
+  .f <- function(...) {
     enquos(...)
   }
 
   quo <- .f(x, y, z)
-  
+
   .serialize <- bdd$serialize(quo)
-  
+
   expect_equal(
-    length(.serialize$args), 
+    length(.serialize$args),
     3
   )
-  
 })
