+++ R/TODO.R
@@ -1,2 +1,0 @@
-# TODO: supress warnings when using columns created during pipeline/
-#       offer more informative message.
