+++ R/apply_transformer.R
@@ -4,8 +4,8 @@ implemented_transforms_base.f <- function(user_defined = list()) {
     "Shuffler" = Shuffler,
     "Encrypter" = Encrypter,
     "Perturber" = Perturber,
-    "Blurer" = Blurer,
-    "NumericBlurer" = NumericBlurer,
+    "Blurrer" = Blurrer,
+    "NumericBlurrer" = NumericBlurrer,
     "GroupedShuffler" = GroupedShuffler,
     "Drop" = Drop,
     "Tidyer" = Tidyer,
@@ -16,8 +16,8 @@ implemented_transforms_base.f <- function(user_defined = list()) {
     "shuffle" = Shuffler,
     "encrypt" = Encrypter,
     "perturb" = Perturber,
-    "blur" = Blurer,
-    "numeric_blur" = NumericBlurer,
+    "blur" = Blurrer,
+    "numeric_blur" = NumericBlurrer,
     "grouped_shuffle" = GroupedShuffler,
     "droper" = Drop,
     "tidy" = Tidyer,
