+++ paper.Rmd
@@ -100,7 +100,7 @@ pipeline4 <- deident(psu, Var3, Var4)
 Performing a multi-stage transformation can be done by chaining together calls to `deident`
 
 ``` {r, eval=F}
-blur <- NumericBlurer$new(cuts=c(0, 10, 100, 1000, 10000))
+blur <- NumericBlurrer$new(cuts=c(0, 10, 100, 1000, 10000))
 
 pipeline3 <- deident(babynames, "psudonymize", name, sex) |>
   deident(blur, n)
