### 2024-12-02 Stat-Cook/deident#14 @Stat-Cook
Purely because neither Blurer or Blurrer appear in the default windows dictionary.  Swapped to blurrer as it appears in the OED.
