+++ NEWS.md
@@ -1,4 +1,4 @@
-# deident 1.0
+# deident 1.0.0
   
 * Implemented a suite of methods for PID removal.
 
