### 2024-12-05 d2670b5
Rename 'Blurer'
