+++ NAMESPACE
@@ -26,11 +26,11 @@ S3method(str,DeidentTask)
 export(":=")
 export(.data)
 export(BaseDeident)
-export(Blurer)
+export(Blurrer)
 export(Drop)
 export(Encrypter)
 export(GroupedShuffler)
-export(NumericBlurer)
+export(NumericBlurrer)
 export(Perturber)
 export(Pseudonymizer)
 export(Shuffler)
