+++ man/Blurrer.Rd
@@ -1,7 +1,7 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/blurer.R
-\name{Blurer}
-\alias{Blurer}
+\name{Blurrer}
+\alias{Blurrer}
 \title{Deidentifier class for applying 'blur' transform}
 \arguments{
 \item{blur}{Look-up list to define aggregation.}
@@ -11,7 +11,7 @@
 \item{...}{Values to be concatenated to keys}
 }
 \value{
-\code{Blurer}
+\code{Blurrer}
 Apply blur to a vector of values
 }
 \description{
@@ -27,6 +27,6 @@ list(London = "England", Paris = "France").
 
 \describe{
 \item{\code{blur}}{List of aggregations to be applied.
-Create new Blurer object}
+Create new Blurrer object}
 }}
 
