+++ man/NumericBlurrer.Rd
@@ -1,7 +1,7 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/numeric_blurer.R
-\name{NumericBlurer}
-\alias{NumericBlurer}
+\name{NumericBlurrer}
+\alias{NumericBlurrer}
 \title{Group numeric data into baskets}
 \description{
 Group numeric data into baskets
