+++ pkgdown.yml
@@ -11,4 +11,4 @@ articles:
   shuffle_example: shuffle_example.html
   transformations: transformations.html
   worked_example: worked_example.html
-last_built: 2024-12-02T16:47Z
+last_built: 2024-12-05T12:17Z
