+++ paper.Rmd
@@ -16,7 +16,7 @@ One approach to dealing with PID concerns is to apply transformations to the dat
 
 In order to broaden the access to sensitive data, data handlers need an auditable open process for the application of data transforms that limit the risk of personal identification (sometimes refereed to as 'data masking' transformations).  Several production scale systems exist for applying data masking transformations (e.g.  [Delphix](https://www.delphix.com/),  [K2View](https://www.k2view.com), and  [Accutive](https://accutive.com/)), but these are often expensive and cumbersome, limiting uptake in the domain of research.  The research community hence requires a simple toolbox for the application of several common 'data masking transforms' implemented  in an open source language.
 
-This implementation of the "deident" methods is in R, chosen due to the increased adoption of open source software into the working practices of the NHS (driven by the growth on the NHS-R community and similar groups).  However, the underlying specification, and implementation of the transforms can be ported to other languages.
+This implementation of the `deident` methods is in R, chosen due to the increased adoption of open source software into the working practices of the NHS (driven by the growth on the NHS-R community and similar groups).  However, the underlying specification, and implementation of the transforms can be ported to other languages.
 
 
 This package was designed considering the ICO guidelines on anonymization  [@ICOAnonymisation] and draws on pre-existing methodologies and naming conventions  [@NIST2015; @integrateIO]. The package implemented de-identification via:   
@@ -42,7 +42,7 @@ siloed within health Trusts.
 
 # Comparison to existing R packages
 
-Several packages have undergone development for the implementation of `encryption` methods to minimize identifiability within data sets (e.g. 'anonymizer'  [@Anonymizer2024], 'deidentifyr'  [@deidentifyr2024] and 'digest' [@digest2021]).  While these packages implement a variety of encryption tools, such systems are not infallible [@wang2005break; @szikora2022end ] especially if the data being encrypted is drawn from a known domain such as common names.  This package introduces the stateful 'pseudonymization' method by which string vectors are replaced by a randomly generated hash the first time they are observed and then preserved for re-use.  As such, breaking a single hash no longer exposes the 'key' and 'salt' the encryption relies on.  In addition, this package introduces methods for removing pseudo-identifiability (identification via the combination of features) via the other methods included.
+Several packages have undergone development for the implementation of encryption methods to minimize identifiability within data sets (e.g. ``anonymizer`  [@Anonymizer2024], `deidentifyr`  [@deidentifyr2024] and `digest` [@digest2021]).  While these packages implement a variety of encryption tools, such systems are not infallible [@wang2005break; @szikora2022end ] especially if the data being encrypted is drawn from a known domain such as common names.  This package introduces the stateful 'pseudonymization' method by which string vectors are replaced by a randomly generated hash the first time they are observed and then preserved for re-use.  As such, breaking a single hash no longer exposes the 'key' and 'salt' the encryption relies on.  In addition, this package introduces methods for removing pseudo-identifiability (identification via the combination of features) via the other methods included.
  
 # In practice
 
