+++ paper.bib
@@ -168,3 +168,12 @@ urldate = {2023-07-08}
      pages = {1686},
      doi = {10.21105/joss.01686},
 }
+
+@Manual{
+    babynames2021,
+    title = {babynames: US Baby Names 1880-2017},
+    author = {Hadley Wickham},
+    year = {2021},
+    note = {R package version 1.0.1},
+    url = {https://CRAN.R-project.org/package=babynames},
+  }
