+++ vignettes/transformations.Rmd
@@ -12,27 +12,21 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
+options(tibble.print_min = 4L, tibble.print_max = 4L)
 ```
 
 Out of the box, `deident` features a set of transformations to aid in the de-identification of data sets.   Each transformation is implemented via `R6Class`  and extends `BaseDeident`.  User defined  transformations can be implemented in a similar manner.
 
-To demonstrate the different transformation we supply a toy data set, `df`, comprising 26 observations of three variables:
+To demonstrate the different transformation we use the `babynames` data set co-opted from the `babynames` package:  
 
-* A:  character, a to z
-* B: numeric, 1 to 26
-* C: character, `X` if `B <= 13`, `Y` if `B > 13`
-
-``` {r, include=F}
+``` {r} 
 library(deident)
 
-df <- data.frame(
-  A = letters,
-  B = 1:26,
-  C = sort(rep(c("X", "Y"), 13))
-)
-df
-```
+babynames <- babynames::babynames |> 
+  dplyr::filter(year > 2015) 
 
+babynames
+```
 ## Psudonymizer
 
 Apply a cached random replacement cipher.  Re-occurrence of the same key will receive the same hash.
@@ -40,20 +34,21 @@ Apply a cached random replacement cipher.  Re-occurrence of the same key will re
 Implemented `deident` options:
 
 ``` {r, eval=F}
-deident(df, "psudonymize", A)
-deident(df, "Pseudonymizer", A)
-deident(df, Pseudonymizer, A)
-deident(df, Pseudonymizer$new(), A)
+# Examples only:
+deident(df, "psudonymize", ...)
+deident(df, "Pseudonymizer", ...)
+deident(df, Pseudonymizer, ...)
+deident(df, Pseudonymizer$new(), ...)
 
 psu <- Pseudonymizer$new()
-deident(df, psu, A)
+deident(df, psu, ...)
 ```
 
 ### Options
 
 By default `Pseudonymizer` replaces  values in variables with a random alpha-numeric string of 5 characters.  This can be replaced via calling `set_method` on an instantiated Pseudonymizer with the desired function:
 
-``` {r}
+``` {r, result="asis"}
 psu <- Pseudonymizer$new()
 
 new_method <- function(key, ...) {
@@ -62,7 +57,10 @@ new_method <- function(key, ...) {
 
 psu$set_method(new_method)
 
-deident(df, psu, A)
+dlist_psu <- deident(babynames, psu, name)
+dlist_psu
+
+apply_deident(babynames, dlist_psu)
 ```
 
 The first argument to the method receives the key to be transformed.
@@ -72,13 +70,14 @@ The first argument to the method receives the key to be transformed.
 Implemented `deident` options:
 
 ``` {r, eval=F}
-deident(df, "shuffle", A)
-deident(df, "Shuffler", A)
-deident(df, Shuffler, A)
-deident(df, Shuffler$new(), A)
+# Examples only:
+deident(df, "shuffle", ...)
+deident(df, "Shuffler", ...)
+deident(df, Shuffler, ...)
+deident(df, Shuffler$new(), ...)
 
 shuffle <- Shuffler$new()
-deident(df, shuffle, A)
+deident(df, shuffle, ...)
 ```
 
 ## Encrypter
@@ -88,13 +87,14 @@ Apply cryptographic hashing to a variable.
 Implemented `deident` options:
 
 ``` {r, eval=F}
-deident(df, "encrypt", A)
[…]
