+++ DESCRIPTION
@@ -25,7 +25,8 @@ Suggests:
     knitr,
     rmarkdown,
     roxygen2,
-    covr
+    covr,
+    babynames
 Imports:
     R6,
     dplyr,
