+++ pkgdown.yml
@@ -11,4 +11,4 @@ articles:
   shuffle_example: shuffle_example.html
   transformations: transformations.html
   worked_example: worked_example.html
-last_built: 2024-12-05T12:42Z
+last_built: 2024-12-05T15:26Z
