### 2024-12-05 Stat-Cook/deident#15 @Stat-Cook
@nrennie  we've updated the Transformations vignette to use the same `babynames` data as the paper (more realistic than the previous `df`).  Is it clear that, for each topic, the first block is the different ways to call the `deident` function, and then the practical examples follow?
