### 2024-12-14 #7157 @spholmes
@Stat-Cook : The review is now completed, to go on to acceptance, I just need you to check the zenodo/figshare references and the items on the above checklist for authors, I will do my checks on my side. Thanks
