+++ NEWS.md
@@ -1,3 +1,9 @@
+# deident 1.1.0
+
+* JOSS review version
+* Rework of vignettes
+* Renaming of 'Blurer' to 'Blurrer'
+
 # deident 1.0.0
   
 * Implemented a suite of methods for PID removal.
