+++ DESCRIPTION
@@ -1,11 +1,11 @@
 Package: deident
 Type: Package
 Title: Persistent Data Anonymization Pipeline
-Version: 1.0.0
+Version: 1.1.0
 Authors@R: c(
     person("Robert", "Cook", , "robert.cook@staffs.ac.uk", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-3343-8271")),
-    person("Md", "Assaduzaman", role = "aut",
+    person("Md", "Assaduzzaman", role = "aut",
            comment = c(ORCID = "0000-0002-8885-6721")),
     person("Sarahjane", "Jones", role = "aut",
            comment = c(ORCID = "0000-0003-4729-4029"))
