+++ paper.bib
@@ -21,7 +21,8 @@
   title={High-tech plans for the NHS},
   author={Wilson, Clare},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/S0262-4079(19)31555-6}
 }
 
 @article{nelson2019predicting,
@@ -32,7 +33,8 @@
   number={1},
   pages={26},
   year={2019},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41746-019-0103-3}
 }
 
 @article{yu2018artificial,
@@ -43,7 +45,8 @@
   number={10},
   pages={719--731},
   year={2018},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41551-018-0305-z}
 }
 
 @book{sales2000ethics,
@@ -74,7 +77,8 @@ urldate = {2023-07-08}
   author={Tachepun, Chitanut and Thammaboosadee, Sotarat},
   booktitle={Proceedings of the 11th international conference on advances in information technology},
   pages={1--9},
-  year={2020}
+  year={2020},
+  doi={10.1145/3406601.3406627}
 }
 
 @Manual{
@@ -110,16 +114,8 @@ urldate = {2023-07-08}
   booktitle={Annual international conference on the theory and applications of cryptographic techniques},
   pages={19--35},
   year={2005},
-  organization={Springer}
-}
-
-@incollection{szikora2022end,
-  title={The end of encryption?--The era of quantum computers},
-  author={Szikora, P{\'e}ter and Laz{\'a}nyi, Korn{\'e}lia},
-  booktitle={Security-Related Advanced Technologies in Critical Infrastructure Protection: Theoretical and Practical Approach},
-  pages={61--72},
-  year={2022},
-  publisher={Springer}
+  organization={Springer},
+  doi={10.1007/11426639_2}
 }
 
 @book{szikora2022end,
@@ -128,7 +124,8 @@ urldate = {2023-07-08}
   booktitle={Security-Related Advanced Technologies in Critical Infrastructure Protection: Theoretical and Practical Approach},
   pages={61--72},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+    doi={10.1007/978-94-024-2174-3_5}
 }
 
 @booklet{ICOAnonymisation,
@@ -145,7 +142,8 @@ urldate = {2023-07-08}
   publisher = {NIST},
   year         = 2015,
   month        = {October},
-  address      = {Gaithersburg, MD}
+  address      = {Gaithersburg, MD},
+  doi = {10.6028/NIST.IR.8053}
 }
 
 
