### 2024-12-18 #7157 @spholmes
@Stat-Cook : Maybe the bot can't find the bibtex file any more?
