### 2025-01-06 #7157 @Stat-Cook
@spholmes Could the problem arise because we are compiling Rmarkdown to markdown which adds citations?  the .md file doesn't have any ref tags for the bot to find?
