+++ paper.md
@@ -28,19 +28,18 @@ date: 12 December 2023
 The delivery of quality health care is a constant act of balancing
 demand against capacity, with emerging, data intensive, artificial
 intelligence (AI) and machine learning (ML) approaches poised to bridge
-gaps in the large, resource-limited sector (Harwich and Laycock 2018;
-Wilson 2019; Nelson et al. 2019; Yu, Beam, and Kohane 2018). The scale
+gaps in the large, resource-limited sector [@harwich2018thinking;
+@wilson2019high; @nelson2019predicting; @yu2018artificial]. The scale
 of data required in such projects magnifies the importance of existing
-ethical and legal frameworks for research with human participants (Sales
-and Folkman 2000; UKRI 2022), notably around the risks posed by the
+ethical and legal frameworks for research with human participants
+[@sales2000ethics; @UKRI], notably around the risks posed by the
 processing of personally identifiable data (PID) and pseudo-PID
-(variables which if used together can identify an individual) (ICO
-2023).
+(variables which if used together can identify an individual) [@ICO].
 
 One approach to dealing with PID concerns is to apply transformations to
 the data, e.g. encryption of names, or aggregation of ages, which can
-limit the risk of identification at the cost of nuance (Tachepun and
-Thammaboosadee 2020). Hence, we demonstrate an extendable package of
+limit the risk of identification at the cost of nuance
+[@tachepun2020data]. Hence, we demonstrate an extendable package of
 tools for the implementation and application of deidentification
 techniques to panel data sets.
 
@@ -65,9 +64,9 @@ groups). However, the underlying specification, and implementation of
 the transforms can be ported to other languages.
 
 This package was designed considering the ICO guidelines on
-anonymization (ICO 2012) and draws on pre-existing methodologies and
-naming conventions (Garfinkel 2015; Integrate.IO, n.d.). The package
-implemented de-identification via:
+anonymization [@ICOAnonymisation] and draws on pre-existing
+methodologies and naming conventions [@NIST2015; @integrateIO]. The
+package implemented de-identification via:
 
 -   pseudonymization – the consistent replacement of a string by a
     random string
@@ -80,27 +79,27 @@ implemented de-identification via:
 -   perturbation – the addition of user-defined random noise to a
     numeric variable
 
-Following the design principles of the existing tidyverse (Wickham et
-al. 2019) domain for ease of adoption. The package includes tools to
-create a single pipeline for the application of a multi-step
-deidentification pipeline to multiple files stored within the same
-repository, alongside the option to serialize/ define the pipeline to/
-via yaml. This approach allows a researcher to design and implement an
-appropriate de-identification plan and deliver it to the research
-support/ business intelligence team of an organisation with limited
-knowledge of the sensitive data. The supply of an easy method for
-de-identifying data sets which requires little scripting knowledge by
-Trust staff may aid in overcoming several information governance risks
-that keep operational data siloed within health Trusts.
+Following the design principles of the existing tidyverse [@tidyverse]
+domain for ease of adoption. The package includes tools to create a
+single pipeline for the application of a multi-step deidentification
[…]
