### 2025-01-06 4c8c06b
Add bibliography to preamble.
