+++ paper.bib
@@ -88,6 +88,7 @@ urldate = {2023-07-08}
     year = {2021},
     note = {R package version 0.6.29},
     url = {https://CRAN.R-project.org/package=digest},
+    doi = {10.32614/cran.package.digest}
   }
 
 @misc{
@@ -174,4 +175,5 @@ urldate = {2023-07-08}
     year = {2021},
     note = {R package version 1.0.1},
     url = {https://CRAN.R-project.org/package=babynames},
+    doi = {10.32614/cran.package.babynames}
   }
