### 2025-01-06 4d58dc1
Fix replace for repeated brackets
