### 2025-01-06 #7157 @Stat-Cook
@spholmes Fixed the referencing and updated for Missing DOIs
