### 2025-01-14 #7157 @spholmes
@Stat-Cook : One last task seems missing to finalize the version and send it on for acceptance and that is
for you to make a new release (if possible) of the software being reviewed and deposit a new archive of  the software with Zenodo/figshare.
Then I will update the review thread with this archive DOI: @editorialbot set 10.5281/zenodo.xxxxxx as archive where .xxxxx is the number you get from zenodo and send me.
