### 2025-01-17 #7157 @spholmes
👋 @crvernon, I think this could go to the next level, I checked the references and DOIs, I think they are now sorted
