### 2025-01-21 #7157 @crvernon
:wave: @Stat-Cook - I just need you to address the following before I move to accept this for publication:

In the archive:

- [ ] The name of your archive should match the name of the paper exactly.  Please edit the archive title to match.
- [ ] Currently, the declared version that you want published is 1.1.0, but your software release and the archive is 1.1.1.  I am assuming that you want 1.1.1 to be the version that is published?

In the paper:

- [ ] LINES 43-45:  I find the use of forward slashes in these sentences awkward.  Please replace with "or" or something similar.  For example, "...the option to serialize or define the pipeline..."
- [ ] LINE 96:  "r" as in the R programming language should be capitalized.  To maintain formatting in your bib file, place characters you wish to maintain formatting of in curly brackets.

No need to make a new release after these changes.  Let me know when they have been addressed and I'll accept this for publication.  Thanks.
