+++ paper.bib
@@ -83,7 +83,7 @@ urldate = {2023-07-08}
 
 @Manual{
     digest2021,
-    title = {digest: Create Compact Hash Digests of R Objects},
+    title = {digest: Create Compact Hash Digests of {R} Objects},
     author = {Dirk Eddelbuettel with contributions by Antoine Lucas and Jarek Tuszynski and Henrik Bengtsson and Simon Urbanek and Mario Frasca and Bryan Lewis and Murray Stokely and Hannes Muehleisen and Duncan Murdoch and Jim Hester and Wush Wu and Qiang Kou and Thierry Onkelinx and Michel Lang and Viliam Simko and Kurt Hornik and Radford Neal and Kendon Bell and Matthew {de Queljoe} and Ion Suruceanu and Bill Denney and Dirk Schumacher and and Winston Chang.},
     year = {2021},
     note = {R package version 0.6.29},
@@ -93,7 +93,7 @@ urldate = {2023-07-08}
 
 @misc{
   Anonymizer2024, 
-  title={Anonymizer: Anonymize data containing personally identifiable information (PII) in R}, 
+  title={Anonymizer: Anonymize data containing personally identifiable information (PII) in {R}}, 
   url={https://github.com/paulhendricks/anonymizer}, 
   author={Hendricks, Paul}, 
   year = {2017},
